(SERVER-2172) Move PuppetConfig class to own file
